Move all component types to the same module.
