+++ libs/pavexc/src/compiler/component/error_handler.rs
