+++ libs/pavexc/src/compiler/component/mod.rs
@@ -1,3 +1,9 @@
+mod constructor;
+mod error_handler;
 mod wrapping_middleware;
+mod request_handler;
 
 pub(crate) use wrapping_middleware::{WrappingMiddleware, WrappingMiddlewareValidationError};
+pub(crate) use constructor::{Constructor, ConstructorValidationError};
+pub(crate) use error_handler::{ErrorHandler, ErrorHandlerValidationError};
+pub(crate) use request_handler::{RequestHandler, RequestHandlerValidationError};
