+++ libs/pavexc/src/compiler/mod.rs
@@ -8,11 +8,8 @@ mod codegen;
 mod codegen_utils;
 mod computation;
 mod component;
-mod constructors;
-mod error_handlers;
 mod generated_app;
 mod interner;
-mod request_handlers;
 mod resolvers;
 mod route_parameter_validation;
 mod traits;
